Rename OrderServiceApplication to CommerceServiceApplication and update related test class
