+++ src/test/java/vn/vinaacademy/order/CommerceServiceApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class OrderServiceApplicationTests {
+class CommerceServiceApplicationTests {
 
     @Test
     void contextLoads() {
